Config: Move config file generation to tabby_config
Keep the models as a separate reference file without any extra
functions.
